pdf-learning-assistant/requirements.txt
HuggingFaceEmbeddings
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-core
cassio
python-dotenv
fastapi
uvicorn
pydantic
requests